move pybind directory and fix header
