+++ cpp/pybind/pipelines/color_map/color_map.h
