+++ cpp/pybind/pipelines/registration/feature.cpp
@@ -28,7 +28,7 @@
 #include "open3d/geometry/PointCloud.h"
 
 #include "pybind/docstring.h"
-#include "pybind/registration/registration.h"
+#include "pybind/pipelines/registration/registration.h"
 
 namespace open3d {
