+++ cpp/pybind/pipelines/registration/registration.cpp
@@ -34,7 +34,7 @@
 #include "open3d/utility/Console.h"
 
 #include "pybind/docstring.h"
-#include "pybind/registration/registration.h"
+#include "pybind/pipelines/registration/registration.h"
 
 namespace open3d {
