+++ cpp/pybind/pipelines/registration/registration.h
